Delete features of questionable utility
- I gave the ax to routing, 16 channel support, and bad graphics.

The 16 channel support in particular was very poorly done, with
immense amount of code duplication and "bool is2ch" parameters.
If it ever comes back, it must be properly designed that time.

Routing seems useless to me with just 2 channels.

I hate the soundflower name and icons. I'll probably think of
something else to replace them.